Fix negative account value problem
Normally, you would deposit cash and then buy stocks (as an example).
Technically, of course, it is possible to buy stocks without having
cash (i.e. margin), or sell stocks that you don't have (sell short),
so it should be possible to have negative balances of accounts. Also,
it should be possible for someone entering a bunch of transactions to
enter them out of order without the app complaining. So, these fixes
allow that to happen.